Move Esent support down to Workspaces.Desktop.  This way we use it from RemoteWorkspace
