+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.AbstractTable.cs
@@ -3,7 +3,7 @@
 using Microsoft.Isam.Esent.Interop;
 using Microsoft.Isam.Esent.Interop.Vista;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
