+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.AbstractTableAccessor.cs
@@ -4,7 +4,7 @@ using System;
 using Microsoft.Isam.Esent.Interop;
 using Roslyn.Utilities;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
