+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.DocumentTable.cs
@@ -1,19 +1,22 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using Microsoft.Isam.Esent.Interop;
 using Roslyn.Utilities;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
-        private class SolutionTable : AbstractTable
+        public class DocumentTable : ProjectDocumentTable
         {
-            private const string TableName = "SolutionTable";
-
+            private const string TableName = "DocumentTable";
             private const string NameColumnName = "Name";
             private const string ValueColumnName = "Value";
 
+            private JET_COLUMNID _projectColumnId;
+            private JET_COLUMNID _projectNameColumnId;
+            private JET_COLUMNID _documentColumnId;
             private JET_COLUMNID _nameColumnId;
             private JET_COLUMNID _valueColumnId;
 
@@ -22,18 +25,20 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
                 var nameColumnCreate = CreateIdColumn(NameColumnName);
                 var valueColumnCreate = CreateBinaryColumn(ValueColumnName);
 
-                var columns = new JET_COLUMNCREATE[] { nameColumnCreate, valueColumnCreate };
+                var columns = CreateProjectDocumentColumns(nameColumnCreate, valueColumnCreate);
 
-                var nameIndexKey = CreateIndexKey(NameColumnName);
+                var primaryIndexKey = CreateProjectDocumentIndexKey(NameColumnName);
 
                 var indexes = new JET_INDEXCREATE[]
                 {
-                    CreatePrimaryIndex(nameIndexKey)
+                    CreatePrimaryIndex(primaryIndexKey)
                 };
 
                 var tableCreate = CreateTable(TableName, columns, indexes);
 
                 Api.JetCreateTableColumnIndex3(sessionId, databaseId, tableCreate);
+
+                GetColumnIds(columns, out _projectColumnId, out _projectNameColumnId, out _documentColumnId);
 
                 _nameColumnId = nameColumnCreate.columnid;
                 _valueColumnId = valueColumnCreate.columnid;
@@ -45,6 +50,8 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
             {
                 using (var table = new Table(sessionId, databaseId, TableName, OpenTableGrbit.ReadOnly))
                 {
+                    GetColumnIds(sessionId, table, out _projectColumnId, out _projectNameColumnId, out _documentColumnId);
+
                     _nameColumnId = Api.GetTableColumnid(sessionId, table, NameColumnName);
                     _valueColumnId = Api.GetTableColumnid(sessionId, table, ValueColumnName);
                 }
@@ -52,7 +59,9 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
 
             public override AbstractTableAccessor GetTableAccessor(OpenSession openSession)
             {
-                return new SolutionTableAccessor(openSession, TableName, PrimaryIndexName, _nameColumnId, _valueColumnId);
+                return new DocumentTableAccessor(
+                    openSession, TableName, PrimaryIndexName,
+                    _projectColumnId, _projectNameColumnId, _documentColumnId, _nameColumnId, _valueColumnId);
             }
         }
     }
