+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.IdentifierLocationTableAccessor.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Threading;
 using Microsoft.Isam.Esent.Interop;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
