+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.OpenSession.cs
@@ -4,7 +4,7 @@ using System;
 using System.Threading;
 using Microsoft.Isam.Esent.Interop;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
