+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.Pool.cs
@@ -3,7 +3,7 @@
 using Microsoft.Isam.Esent.Interop;
 using Roslyn.Utilities;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
