+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.ProjectDocumentTable.cs
@@ -2,7 +2,7 @@
 
 using Microsoft.Isam.Esent.Interop;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
