+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.SolutionTable.cs
@@ -1,22 +1,19 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using Microsoft.Isam.Esent.Interop;
 using Roslyn.Utilities;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
-        public class DocumentTable : ProjectDocumentTable
+        private class SolutionTable : AbstractTable
         {
-            private const string TableName = "DocumentTable";
+            private const string TableName = "SolutionTable";
+
             private const string NameColumnName = "Name";
             private const string ValueColumnName = "Value";
 
-            private JET_COLUMNID _projectColumnId;
-            private JET_COLUMNID _projectNameColumnId;
-            private JET_COLUMNID _documentColumnId;
             private JET_COLUMNID _nameColumnId;
             private JET_COLUMNID _valueColumnId;
 
@@ -25,20 +22,18 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
                 var nameColumnCreate = CreateIdColumn(NameColumnName);
                 var valueColumnCreate = CreateBinaryColumn(ValueColumnName);
 
-                var columns = CreateProjectDocumentColumns(nameColumnCreate, valueColumnCreate);
+                var columns = new JET_COLUMNCREATE[] { nameColumnCreate, valueColumnCreate };
 
-                var primaryIndexKey = CreateProjectDocumentIndexKey(NameColumnName);
+                var nameIndexKey = CreateIndexKey(NameColumnName);
 
                 var indexes = new JET_INDEXCREATE[]
                 {
-                    CreatePrimaryIndex(primaryIndexKey)
+                    CreatePrimaryIndex(nameIndexKey)
                 };
 
                 var tableCreate = CreateTable(TableName, columns, indexes);
 
                 Api.JetCreateTableColumnIndex3(sessionId, databaseId, tableCreate);
-
-                GetColumnIds(columns, out _projectColumnId, out _projectNameColumnId, out _documentColumnId);
 
                 _nameColumnId = nameColumnCreate.columnid;
                 _valueColumnId = valueColumnCreate.columnid;
@@ -50,8 +45,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
             {
                 using (var table = new Table(sessionId, databaseId, TableName, OpenTableGrbit.ReadOnly))
                 {
-                    GetColumnIds(sessionId, table, out _projectColumnId, out _projectNameColumnId, out _documentColumnId);
-
                     _nameColumnId = Api.GetTableColumnid(sessionId, table, NameColumnName);
                     _valueColumnId = Api.GetTableColumnid(sessionId, table, ValueColumnName);
                 }
@@ -59,9 +52,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
 
             public override AbstractTableAccessor GetTableAccessor(OpenSession openSession)
             {
-                return new DocumentTableAccessor(
-                    openSession, TableName, PrimaryIndexName,
-                    _projectColumnId, _projectNameColumnId, _documentColumnId, _nameColumnId, _valueColumnId);
+                return new SolutionTableAccessor(openSession, TableName, PrimaryIndexName, _nameColumnId, _valueColumnId);
             }
         }
     }
