+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.StringNameTableAccessor.cs
@@ -3,7 +3,7 @@
 using System.Text;
 using Microsoft.Isam.Esent.Interop;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
