+++ src/Workspaces/Core/Desktop/Workspace/Esent/EsentStorage.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using Microsoft.Isam.Esent.Interop;
 using Roslyn.Utilities;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation.Esent
+namespace Microsoft.CodeAnalysis.Esent
 {
     internal partial class EsentStorage
     {
