+++ src/Workspaces/Core/Desktop/Workspaces.Desktop.csproj
@@ -66,6 +66,28 @@
       <DependentUpon>WorkspaceDesktopResources.resx</DependentUpon>
     </Compile>
     <Compile Include="Workspace\CommandLineProject.cs" />
+    <Compile Include="Workspace\Esent\EsentLogger.cs" />
+    <Compile Include="Workspace\Esent\EsentPersistentStorage.cs" />
+    <Compile Include="Workspace\Esent\EsentPersistentStorage_IdentifierTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.AbstractTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.AbstractTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.DocumentTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.DocumentTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.IdentifierLocationTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.IdentifierLocationTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.IdentifierTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.NameTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.OpenSession.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.Pool.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.ProjectDocumentTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.ProjectDocumentTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.ProjectTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.ProjectTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.SolutionTable.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.SolutionTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.StringNameTableAccessor.cs" />
+    <Compile Include="Workspace\Esent\EsentStorage.TableKinds.cs" />
     <Compile Include="Workspace\FileTextLoader.cs" />
     <Compile Include="Workspace\Host\Documentation\DocumentationProviderServiceFactory.cs" />
     <Compile Include="Workspace\Host\Mef\DesktopMefHostServices.cs" />
@@ -95,6 +117,8 @@
     <Compile Include="Workspace\MSBuild\VisualBasic\VisualBasicProjectFileLoader.cs" />
     <Compile Include="Workspace\MSBuild\VisualBasic\VisualBasicProjectFileLoaderFactory.cs" />
     <Compile Include="Workspace\MSBuild\MSBuildProjectLoader.cs" />
+    <Compile Include="Workspace\SolutionSize\SolutionSizeTracker.cs" />
+    <Compile Include="Workspace\Storage\PersistentStorageService.cs" />
   </ItemGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="Microsoft.CodeAnalysis.CSharp.EditorFeatures" />
