+++ src/Workspaces/Core/Portable/Workspace/Host/PersistentStorage/IPersistentStorageService.cs
