+++ src/Workspaces/Core/Portable/Workspace/Host/PersistentStorage/PersistentStorageOptions.cs
@@ -9,5 +9,7 @@ namespace Microsoft.CodeAnalysis.Host
         public const string OptionName = "FeatureManager/Persistence";
 
         public static readonly Option<bool> Enabled = new Option<bool>(OptionName, "Enabled", defaultValue: true);
+
+        public static readonly Option<bool> EsentPerformanceMonitor = new Option<bool>(OptionName, "Esent PerfMon", defaultValue: false);
     }
 }
